Reorganize skills into self-describing skill directories
Each skill is now a directory containing skill.md (YAML frontmatter +
description) and tool.py (implementation + factory). A shared registry
in __init__.py exposes collect_tools() and _base.py provides load_skill_meta()
and slugify() utilities.

main.py now calls collect_tools() instead of importing hardcoded tool lists.

Deleted flat modules: base.py, search.py, library.py, memory.py,
file_write.py, steps.py, and the docs/ directory.
