+++ bartleby/write/skills/get_full_document/skill.md
@@ -1,3 +1,21 @@
+---
+name: get_full_document
+agents: [search_expert]
+inputs:
+  document_id:
+    type: string
+    description: "Document ID from search results"
+  start_chunk:
+    type: integer
+    description: "Zero-based index to start from (default 0)"
+    nullable: true
+  max_chunks:
+    type: integer
+    description: "Max chunks to return (capped at 100)"
+    nullable: true
+output_type: string
+---
+
 Retrieve chunks from a specific document with pagination.
 
 Use this to read through a document sequentially. Provide the `document_id` from a search result or document listing. Use `start_chunk` to page through long documents.
