+++ bartleby/write/skills/read_notes/skill.md
@@ -1,0 +1,13 @@
+---
+name: read_notes
+agents: [search_expert, research]
+inputs: {}
+output_type: string
+---
+
+Read all saved research notes. Notes are shared between you and other agents, and persist across sessions. Use this to review what has been discovered so far.
+
+**When to use:**
+- At the start of research to check for prior findings
+- When the user asks about previous research
+- To see what the search agent has flagged as important
